+++ routes/v1/user.js
@@ -2,9 +2,9 @@ const express = require('express');
 
 const { imageUpload } = require('../../utils/file');
 const asyncWrap = require('../../utils/errors/wrap');
-const userValid = require('../../utils/validators/user');
+const userValid = require('../../middlewares/validators/user');
 const userController = require('../../controllers/user');
-const studyValid = require('../../utils/validators/study');
+const studyValid = require('../../middlewares/validators/study');
 const studyController = require('../../controllers/study');
 
 const router = express.Router();
@@ -20,7 +20,11 @@ router.patch('/:id', imageUpload, userValid.userUpdate, asyncWrap(userController
 router.delete('/:id', userValid.withdraw, asyncWrap(userController.withdraw)); // 회원 탈퇴
 
 router.get('/emailVerify/:email', userValid.emailVerification, asyncWrap(userController.emailVerification)); // 이메일 인증 요청
-router.get('/emailVerify-process/:email', userValid.emailVerificationProcess, asyncWrap(userController.emailVerificationProcess)); // 이메일 인증 요청
+router.get(
+  '/emailVerify-process/:email',
+  userValid.emailVerificationProcess,
+  asyncWrap(userController.emailVerificationProcess)
+); // 이메일 인증 요청
 
 router.get('/:id/study', studyValid.myStudy, asyncWrap(studyController.myStudy)); // 내 스터디 조회
